refactor: fix typos Co-authored-by: teenager-ETH <sunsets01.bombast@icloud.com>
